requirements.txt
asgiref==3.2.10
Django==3.0.8
django-environ==0.4.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
mysqlclient==2.1.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0
PyJWT==2.4.0
requests==2.28.1